Add more constants.
base_JSON_path - the base path for all JSON files excluding the config as that is a special file.
DB_indent_level - the number of spaces the JSON handler will indent the files with. Useful for adjusting RAM usage.
valid_blacklists - a list of valid blacklists that the user can supply.